+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v0.8.0 (29 Aug 2021)
 
-- fix and improve types.
+- fix and improve types with last kernel types.
 - update deps
 
 ### v0.7.0 (22 Aug 2021)